+++ lib/mock_video.py
@@ -1,0 +1,3 @@
+# For testing the analyzer server
+if __name__ == "__main__":
+    print("bluh");
